Creating, Writing and DeletingHandler finished, polishing new database structure

+++ server/src/main/java/tk/chaber/sfn2021rest/db/UsersRepo.java
@@ -12,5 +12,5 @@ public interface UsersRepo extends CrudRepository<User, Integer> {
 
     List<User> findByToken(byte[] token);
 
-    List<User> findById(int id);
+    List<User> findById(Long id);
 }
